Сократил объём потребляемой каскадом памяти в 25 раз за счёт сохранения всех губок в одном массиве. Добавил аварийную деаллокацию памяти и повторную перезапись значений памяти в аллокаторе Добавил аварийную деаллокацию при завершении программы и усилил надёжность деаллокации Другие мелкие изменения

+++ src/main/4 utils/Memory/Utils-alignment.cs
@@ -10,7 +10,7 @@ public unsafe static partial class Utils
     /// <param name="size">Размер массива для выравнивания</param>
     /// <param name="alignment">Размер границ, на который выравнивается</param>
     /// <returns>Выравненное значение size</returns>
-    public static int calcAlignment(int size, int alignment = 64)
+    public static nint calcAlignment(nint size, int alignment = 64)
     {
         var bmod = size % alignment;
         if (bmod == 0)
